Refactoring and bug fixes
Haven't got any unit tests for SiteIndexer or ContentIndexer because I was
finding it hard to think through what I need to mock to write a test. I think
just the hostSetting and getSiteContent (get it to return an iterator with a
fake query that resembles some mura content). I'll make this something to do
later, for now I've just tested the reindex, update, and remove functionality
by hand.