Updated admin layout
I think this is close to final now, just going to make small improvements and a
few additional features need to be implemented then I'm going to consider the
interface done. Might be a good idea to create some tests but for now I'm not
going to bother because it's easy for me to check that things work myself quickly
and I don't expect I'll make a lot of changes here, and if I do I don't care.