Tweaks to SiteIndexer
Now it will get all content not just live content - this is so that
we can get content that is set to display but not until some time in
the future. The read alias will get a search filter added to it that
will prevent content that should only be visible between a specific
start and end date from displaying leaking into the search results
when it shouldn't. Still not indexing draft versions by default so
this search is not appropriate to replace the back end search, but it
could be made to so if desired. To do this you would need to change
the getDefaultSiteContentFeed to to return unapproved content and then
you would probably want to create a new read alias for the public
searches that filters out content that shouldn't be displayed because
the displayDate is not right and because it is a draft.

I am also going to simplify the search in the back end so it is just
a search field and a results list without any filtering as it's
too much work for too little value to do more coming back to this!
You can take a look at https://github.com/oscarduignan/react-rxjs-elasticsearch-faceted-search-example
for an example of a more advanced interface.