Updated the admin search
Simplified it, removed filters, got query working - can now do something
like "30 OR 45 -lorem" to get anything with 30 or 45 in it that doesn't
have the word -lorem anywhere in the fields we're searching. Will refactor
it more in the future, eventually would be nice to strip out all the reflux
stuff that is confusing and move to using rxjs like I've done with
https://github.com/oscarduignan/react-rxjs-elasticsearch-faceted-search-example
but that isn't a big priority!

Stuff I've noticed to do:

- the refresh of indexing status seems to happen to quickly after you click loading,
  I think it should optimistically create an entry that says "loading" so you know
  that the click of reindex all has done something and then it should do something
  like setTimeout to manage the refreshing every 10 seconds (maybe do one 5 seconds
  after the click to reindex).

- appreload locally for me is crashing lucee, might be unrelated to the plugin but
  needs some investigating.

- add pagination to the search results in the admin.

- create a display object to replace the default search.

- create an advanced display object with some filters.

- more test coverage, followed by some refactoring to tidy things up.